Fixes sample plugin file
The documentation here 
https://github.com/phonegap/phonegap-docs/blob/master/docs/2-tutorials/2-develop/1-embed-webview/1-ios.html.md

references this sample file.  However this sample file is invalid and does not match the sample podfile embedded inline in the documentation.  

The reason this sample file is invalid because `cordova-plugin-wkwebview-engine` requires ios 9 as a minimum target platform and this file specifies ios8.  According to the link above this pod isn't required at all.